mmc: core: Do a power cycle when the CMD11 fails
PROPAGATED_from (CR)

A CMD11 is sent to the SD/SDIO card to start the voltage switch procedure
into 1.8V I/O. According to the SD spec a power cycle is needed of the
card, if it turns out that the CMD11 fails. Let's fix this, to allow a
retry of the initialization without the voltage switch, to succeed.

Note that, whether it makes sense to also retry with the voltage switch
after the power cycle is a bit more difficult to know. At this point, we
treat it like the CMD11 isn't supported and therefore we skip it when
retrying.

Mot-CRs-Fixed: (CR)

Change-Id: I6fb94573f330c3c9bca0d6266a8f0f1f07d84e77
Signed-off-by: DooHyun Hwang <dh0421.hwang@samsung.com>
Link: https://lore.kernel.org/r/20210210045936.7809-1-dh0421.hwang@samsung.com
Cc: stable@vger.kernel.org
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Reviewed-on: https://gerrit.mot.com/2035336
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Wang Wang <wangwang1@lenovo.com>
Reviewed-by: Shengchun Li <lisc2@motorola.com>
Submit-Approved: Jira Key
Reviewed-on: https://gerrit.mot.com/2038932
Reviewed-by: Qiong Xie <xieq2@motorola.com>
Reviewed-on: https://gerrit.mot.com/2192540
Reviewed-by: Jianguo Liu <liujg6@motorola.com>
Reviewed-by: Robert Cabral <robertc@motorola.com>